requirements.txt
openai==0.28.0
pandas==2.2.2
langchain==0.2.14
langchain_community==0.2.12
langchain-huggingface==0.0.3
gradio==4.42.0
transformers==4.44.2
sentence-transformers==3.0.1
faiss-cpu==1.8.0
python-dotenv==1.0.0
